tests: Move test_auto_subbed_to_personals test to test_subs.py.
